Add files via upload
This is a class java that extend the understanding of object and its relationship with other classes. For instance, A given class name Car posses many other part that helps to for the car's structure as the engine, the internal items and so on. The engine could be programmed as an class that describe the engine that a particular car has, then on the class Car, it is possible to set in the parameters variable a new one as a type as engine. This in simple term describe the relationship between different classes in Java.