Interface
These are a more sophisticated example of interface where we created two interface (Predator and Prey). These two interface will have different methods that must be implement by the classes that want to implement from the interface. Then, the classes Fish, Lion and Buffalo must have the methods from the respective implementation add internally.  These examples are the best when it comes to understanding the use of interfaces in Java.