Exceptions
This is a simple class java that illustrate the user of exception and how to apply it in simple code. Firstly, exception may vary in type, but almost every code may have its need to treat exception or issues that occur during the code execution. Exception can be implemented as a way to obtain the error origin and then solve it by analyse the error.