String methods
This is a example of a class that shows the many methods available in the class String. It covers the construction of a string, the localization process for an specific character, the comparation between one or more character or strings, the extraction of an particular letter within the string and the modification process which would be suitable if we want to let all the character as a upper or lowercase.